+++ src/pydvl/utils/parallel/backend.py
@@ -136,9 +136,8 @@ class RayParallelBackend(BaseParallelBackend, backend_name="ray"):
         config_dict.pop("backend")
         config_dict["num_cpus"] = config_dict.pop("n_local_workers")
         self.config = config_dict
-        if self.config["address"] is None:
-            self.config["ignore_reinit_error"] = True
-        ray.init(**self.config)
+        if not ray.is_initialized():
+            ray.init(**self.config)
 
     def get(
         self,
@@ -200,7 +199,7 @@ def init_parallel_backend(
     >>> config = ParallelConfig(backend="ray")
     >>> parallel_backend = init_parallel_backend(config)
     >>> parallel_backend
-    <RayParallelBackend: {'address': None, 'logging_level': 30, 'num_cpus': None, 'ignore_reinit_error': True}>
+    <RayParallelBackend: {'address': None, 'logging_level': 30, 'num_cpus': None}>
 
     """
     try: